requirements.txt
mssql-django
python-dotenv
pillow